+++ template/base-ui/src/main/java/ru/surfstudio/standard/base_ui/loadstate/presentation/NoneLoadStatePresentation.kt
@@ -1,6 +1,6 @@
 package ru.surfstudio.standard.base_ui.loadstate.presentation
 
-import ru.surfstudio.standard.base_ui.loadstate.PlaceHolderViewContainer
+import ru.surfstudio.android.custom.view.placeholder.PlaceHolderViewContainer
 import ru.surfstudio.android.core.mvp.loadstate.SimpleLoadStatePresentation
 import ru.surfstudio.standard.base_ui.loadstate.state.NoneState